move reporting to its own slot instead of being called from run_tests
This separates out reporting to its own callable `def report()` on Mozmill.
I've cleaned up the control flow a bit.  While the patch is pretty pervasive,
it (basically) doesn't affect what is done, and (mostly) only affects when
things are done for the case of reporting.  In the future, the control flow
should be revisited, probably in conjunction with or after the CLI cleanup.

* fixed with review comments from ahal

* add the firePythonCallback listener to MozmillRestart start()

* add a hard kill option for fatal errors

* document get_report method

* factor out reporting to its own method

* make the CLI handling code simpler...its all worth it

* unify stop and stop_runner functions and have stop call stop_runner

* dont bothering raising the TestsFailedException at all since the only result
is sys.exit; just call sys.exit

* set self.profile in the right place so that the runner can be created without
a ValueError

* removed unused variable "report" from run_dir

* document the behaviour of run_dir

* document the 1-run nature of MozMill

* remove yet another call to super

* output to stdout like everything else instead of stderr

* document run_tests method

* remove unused variable self.report_document

* note poor control flow for Try Except

* move the try: except: into the inner function and eliminate _run entirely;
also get rid of the silly warning message for MozmillRestarts CLI class