Updated two README.md to remove dead addon link.
The addon link on the pages leads to a page that says: "This add-on has
been removed by its author.". Furthermore, by searching on Google I was
able to find the following Mozmill Developpers Google Groups message
(https://groups.google.com/forum/#!topic/mozmill-dev/02HwCHWQFRA) that
states the addon was removed because it hadn't been updated in a long
time and support for it had already been dropped.

Corrected another README.md